+++ Potraviny/src/main/java/Ovocie/Ananas.java
@@ -16,11 +16,12 @@ public class Ananas extends Ovocie_Zelenina {
 
     private Farba farba;
 
-    public Ananas(Farba farba, String nazov, double hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
+    public Ananas(Farba farba, String nazov, int hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
         super(nazov, hmotnost, cena, sladke, supka, kysele);
         this.farba = farba;
     }
 
+   
    
     public Farba getFarba() {
         return farba;
@@ -40,11 +41,11 @@ public class Ananas extends Ovocie_Zelenina {
         this.nazov = nazov;
     }
 
-    public double getHmotnost() {
+    public int getHmotnost() {
         return hmotnost;
     }
 
-    public void setHmotnost(double hmotnost) {
+    public void setHmotnost(int hmotnost) {
         this.hmotnost = hmotnost;
     }
