+++ Potraviny/src/main/java/Ovocie/Banany.java
@@ -15,12 +15,11 @@ import CerstveJedlo.Ovocie_Zelenina;
 public class Banany extends Ovocie_Zelenina {
     private Farba farba;
 
-    public Banany(Farba farba, String nazov, double hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
+    public Banany(Farba farba, String nazov, int hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
         super(nazov, hmotnost, cena, sladke, supka, kysele);
         this.farba = farba;
     }
 
-   
     public Farba getFarba() {
         return farba;
     }
@@ -41,7 +40,7 @@ public class Banany extends Ovocie_Zelenina {
         return hmotnost;
     }
 
-    public void setHmotnost(double hmotnost) {
+    public void setHmotnost(int hmotnost) {
         this.hmotnost = hmotnost;
     }
