+++ Potraviny/src/main/java/Ovocie/Citrony.java
@@ -15,11 +15,12 @@ import CerstveJedlo.Ovocie_Zelenina;
 public class Citrony extends Ovocie_Zelenina {
     private Farba farba;
 
-    public Citrony(Farba farba, String nazov, double hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
+    public Citrony(Farba farba, String nazov, int hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
         super(nazov, hmotnost, cena, sladke, supka, kysele);
         this.farba = farba;
     }
-   
+
+    
 
     public Farba getFarba() {
         return farba;
@@ -37,11 +38,11 @@ public class Citrony extends Ovocie_Zelenina {
         this.nazov = nazov;
     }
 
-    public double getHmotnost() {
+    public int getHmotnost() {
         return hmotnost;
     }
 
-    public void setHmotnost(double hmotnost) {
+    public void setHmotnost(int hmotnost) {
         this.hmotnost = hmotnost;
     }
