+++ Potraviny/src/main/java/Zelenina/Petrzlen.java
@@ -16,7 +16,7 @@ public class Petrzlen extends Ovocie_Zelenina {
     private Farba farba;
     private boolean korenovaZelenina;
 
-    public Petrzlen(Farba farba, boolean korenovaZelenina, String nazov, double hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
+    public Petrzlen(Farba farba, boolean korenovaZelenina, String nazov, int hmotnost, int cena, boolean sladke, boolean supka, boolean kysele) {
         super(nazov, hmotnost, cena, sladke, supka, kysele);
         this.farba = farba;
         this.korenovaZelenina = korenovaZelenina;
@@ -46,11 +46,11 @@ public class Petrzlen extends Ovocie_Zelenina {
         this.nazov = nazov;
     }
 
-    public double getHmotnost() {
+    public int getHmotnost() {
         return hmotnost;
     }
 
-    public void setHmotnost(double hmotnost) {
+    public void setHmotnost(int hmotnost) {
         this.hmotnost = hmotnost;
     }
